tutorial/chapters/ch5_enhanced_batchapp/requirements.txt
pycarol[complete]
python-dotenv
python-json-logger
sklearn